refactor: extract AI translation logic and centralize type definitions for better modularity
