+++ extensions/bangumi/src/components/MyCollection.tsx
@@ -2,7 +2,8 @@ import { ActionPanel, List, Action, Icon, getPreferenceValues } from "@raycast/a
 import { usePromise } from "@raycast/utils"
 import { useRef, useState } from "react"
 import { bangumi } from "@/api/bangumi"
-import { SubjectCollectionType, SubjectType, SubjectTypeName, getCollectionTag } from "@/shared/const"
+import { SubjectCollectionType, SubjectType, SubjectTypeName } from "@/shared/const"
+import { getCollectionTag } from "@/shared/utils"
 import ProgressViewer from "./ProgressViewer"
 import SubjectDetail from "./SubjectDetail"
 import { CollectionStatusActions, OpenInBgmBrowser } from "./actions"
