+++ extensions/bangumi/src/components/actions/index.ts
@@ -1,3 +1,4 @@
 export * from "./EpisodeStatusActions"
 export * from "./CollectionStatusActions"
 export * from "./OpenInBgmBrowser"
+export * from "./AITranslateAction"
