+++ extensions/bangumi/src/search-characters.tsx
@@ -1,7 +1,8 @@
 import { Action, ActionPanel, Icon, LaunchProps, Grid } from "@raycast/api"
 import { usePromise, withAccessToken } from "@raycast/utils"
 import { useRef, useState } from "react"
-import { bangumi, Infobox } from "@/api/bangumi"
+import { bangumi } from "@/api/bangumi"
+import { Infobox } from "@/types"
 import { bangumiAuth } from "@/api/oauth"
 import { OpenInBgmBrowser } from "@/components/actions"
 import CharacterDetail from "@/components/CharacterDetail"
