+++ extensions/bangumi/src/tools/get-subject-details.ts
@@ -2,7 +2,7 @@ import { bangumi } from "@/api/bangumi"
 import { withAccessToken } from "@raycast/utils"
 import { bangumiAuth } from "@/api/oauth"
 import { formatSubjectToMarkdown, formatInfoboxToMarkdown } from "./utils"
-import { getCollectionTag } from "@/shared/const"
+import { getCollectionTag } from "@/shared/utils"
 
 type Input = {
   /**
