+++ extensions/bangumi/src/tools/update-collection.ts
@@ -1,6 +1,6 @@
 import { Tool } from "@raycast/api"
 import { bangumi } from "@/api/bangumi"
-import { getCollectionTag } from "@/shared/const"
+import { getCollectionTag } from "@/shared/utils"
 import { withAccessToken } from "@raycast/utils"
 import { bangumiAuth } from "@/api/oauth"
