+++ extensions/bangumi/src/types/index.ts
@@ -1,0 +1,9 @@
+import { Color } from "@raycast/api"
+import type { components } from "./generated"
+
+export type Infobox = components["schemas"]["WikiV0"]
+
+export interface CollectionTag {
+  value: string
+  color: Color | string
+}
